requirements.txt
pylint
pyyaml
selenium